+++ src/components/TabMenuComponent/containers/index.ts
@@ -1,0 +1,4 @@
+export * from './HorizontalTabMenuContainer';
+export * from './IconTabMenuContainer';
+export * from './HorizontalTabsContainer';
+export * from './VerticalTabMenuContainer';