+++ src/components/TabMenuComponent/stories/horizontalTabsStories/HorizontalTabs.stories.tsx
@@ -1,0 +1,90 @@
+import styled from 'styled-components';
+import type { Meta, StoryFn } from '@storybook/react';
+import { useGlobals } from '@storybook/preview-api';
+
+import { ALL_BORDER_RADIUS_VALUES, HorizontalTabs } from '@admiral-ds/react-ui';
+
+import { HorizontalTabsTemplate } from './HorizontalTabsTemplate';
+
+// Imports of text sources
+import HorizontalTabsRaw from './HorizontalTabsTemplate?raw';
+
+const Desc = styled.div`
+  font-family: 'VTB Group UI';
+  font-size: 16px;
+  line-height: 24px;
+`;
+
+const Description = () => (
+  <Desc>
+    Компонент настроен таким образом, что если закладки не помещаются в ширину экрана, то они “выходят” за область
+    экрана и их можно прокручивать свайпом, в остальном, поведение такое же, как и у настольной версии. Рекомендуется
+    использовать для мобильной версии.
+  </Desc>
+);
+
+export default {
+  title: 'Admiral-2.1/TabMenuComponent/HorizontalTabs',
+  decorators: undefined,
+  component: HorizontalTabs,
+  parameters: {
+    docs: {
+      source: {
+        code: null,
+      },
+    },
+    componentSubtitle: <Description />,
+  },
+  argTypes: {
+    dimension: {
+      options: ['l', 'm'],
+      control: { type: 'radio' },
+    },
+    showUnderline: {
+      control: { type: 'boolean' },
+    },
+    themeBorderKind: {
+      options: ALL_BORDER_RADIUS_VALUES,
+      control: { type: 'radio' },
+    },
+    defaultSelectedTabId: {
+      control: false,
+    },
+    renderTab: {
+      control: false,
+    },
+    tabIsDisabled: {
+      control: false,
+    },
+    tabsId: {
+      control: false,
+    },
+    onSelectTab: {
+      control: false,
+    },
+    selectedTabId: {
+      control: false,
+    },
+  },
+} as Meta<typeof HorizontalTabs>;
+
+//<editor-fold desc="HorizontalTabs.">
+const HorizontalTabsStory: StoryFn<typeof HorizontalTabs> = (props) => {
+  const [{ CSSCustomProps }] = useGlobals();
+  return <HorizontalTabsTemplate {...props} CSSCustomProps={CSSCustomProps} />;
+};
+
+export const HorizontalTabsExample = {
+  render: HorizontalTabsStory,
+
+  parameters: {
+    docs: {
+      source: {
+        code: HorizontalTabsRaw,
+      },
+    },
+  },
+
+  name: 'HorizontalTabs.',
+};
+//</editor-fold>