+++ src/components/TabMenuComponent/tabElements/TabText.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const TabText = styled.div`
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+`;
+TabText.displayName = 'TabText';