+++ src/components/TabMenuComponent/tabElements/index.ts
@@ -1,0 +1,3 @@
+export * from './TabBadge';
+export * from './TabIcon';
+export * from './TabText';